+++ src/Changes/Manual/RenegadeVikingsChanges/Archetypes/IArchetype.cs
@@ -1,8 +1,0 @@
-using System.Collections.Generic;
-
-namespace LotusEcarlateChanges.Changes.Manual.RenegadeVikingsChanges.Archetypes;
-
-public interface IArchetype
-{
-  public List<CharacterDrop.Drop> Drops { get; }
-}